♻️ refactor: move §2.5 palette to its own file (ADR-028 slice 4 stage 0)
Pure move, no value changes. §2.5's light half leaves `DesignTokens.swift`
(139-180) and its §2.9 dark counterparts leave `DesignTokens+NightPalette.swift`
(231-371), together with the light-only provenance note that described them.

The axis is **a completed section moves out whole** — a third axis after slice
2's "the night data leaves the day file" and the value-vs-invariant one slice 3
declined. §2.5 was finished by slice 3, so unlike every other section it will
not grow again, and taking both halves at once frees headroom in two files from
one move:

  DesignTokens.swift             396 -> 354   (400-line file_length cap)
  DesignTokens+NightPalette.swift 372 -> 238
  DesignTokens+AvatarPalette.swift  new 216

`DesignTokens.swift` was the binding constraint: six of slice 4's ten light
tokens live in it and it had four lines of headroom.

Byte-exactness is verified rather than assumed — all 34 avatar token
declarations resolve in the new file, none are left behind, no non-avatar
declaration moved, and both moved blocks appear contiguous and byte-identical.

Two header claims the move falsified are corrected in place: `+NightPalette`'s
"every `night*` raw token, in one file" and its "this file is where the `night*`
hex literals now live". Both new files keep the `DesignTokens` prefix, without
which `check_design_tokens_css.py`'s glob stops reaching the hexes and the
`tokens.css` mirror gate goes blind with no signal.

See #1325
